pyproject.toml
[project]
name = "adk-cc"
version = "0.1.0"
description = "Minimal Claude-Code-style gather/act/verify agent loop on Google ADK"
requires-python = ">=3.12"
dependencies = [
    "google-adk==1.31.1",
    "litellm>=1.50",
    "filelock>=3.13",
]

[tool.setuptools.packages.find]
include = ["adk_cc*"]
